Implement --document-root and --passthru
